Extract EmployeeHierarchyService from EmployeeService
EmployeeService, the largest of the services, mixed employee CRUD/join/role
concerns with the reporting-hierarchy logic. Move the hierarchy cluster
(assign/remove manager, list subordinates, list managers, and the manager
validation with its circular-reference check) into a focused
EmployeeHierarchyService.

EmployeeService injects it: the five public hierarchy methods become thin
delegates so EmployeeControllerWeb is unchanged, and the patch path calls
employeeHierarchyService.validateManager for a manager change. Behaviour and
the REST surface are unchanged.
